fix(stacktrace): Do not rely on AbsPath of files
NewStacktrace and ExtractStacktrace (through filterFrames) used to
whitelist stack frames of test and example files that are part of
sentry-go. That is likely only done to allow for testing.

However, by relying on the absolute path of files to perform the
filtering, the correct filtering of frames depended on how sentry-go was
checked out on disk. If the repository was checked out as part of a
traditional GOPATH tree, tests passed. If the repository was outside of
GOPATH, on an arbitraly named path (without a github.com/getsentry/
prefix), tests failed.

This commit replaces tests from stacktrace_test.go with new tests in
stacktrace_external_test.go. The difference being that the new tests are
built as a separate package. The separation is achieved by using the
ability to suffix the test package name with '_test'. See 'go help
test'.

This new test structure makes it easier to distinguish stack frames that
belong to the sentry-go SDK vs. stack frames that are part of the tests,
thus removing the need for path-pattern-based filtering.

The new tests make a more in-depth comparison of the stack frames
against expected values.

Now NewStacktrace and ExtractStacktrace filter out all the frames that
are part of the github.com/getsentry/sentry-go package/Module, without
looking at AbsPath.