fix: Report usage error including stack trace
Prior to this, calling CaptureException, Recover or RecoverWithContext
with a nil error would send a message event to Sentry without a stack
trace.

Looking at the Sentry issues list, and issue detail, all you see is
"Called CaptureException with nil value" without a clue of what caused
it.

This commit turns that message into a sentry.usageError with a stack
trace, making it easier to understand what is wrong and fix user code.