ci: Ignore failures when Go Modules are disabled
Without vendoring and without Go Modules, we're testing the master
branch of each dependency which is a combination of moving targets.

Most times it is okay when those cases fail. Integrations like Iris
require use of Go Modules, and will fail to even fetch dependencies when
modules are disabled.

This is the quick fix, a more complete change would be ignoring fetching
dependencies and not testing Iris or any other parts that require Go
Modules when GO111MODULE=off.