fix: Recover* report errors for any type of panic
The previous behavior was to silently stop panics without reporting to
Sentry when the return value from recover() was not an error or a
string.

Runtime panics are specified to implement the predefined error
interface, and string is probably the second most common type passed to
panic() calls, still calls like panic(42) should not pass silently.

This fix converts any other type to string using Go's standard
formatting rules and reports the message to Sentry.