misc: Rename fields and types for tracing support
1. Rename Event.StartTimestamp, Span.StartTimestamp and
Span.EndTimestamp to remove the "stamp" suffix in preparation to use
StartTime and EndTime for spans when full tracing support is added.

Preferring the shorter name because it is shorter and
equally clear, and the type is time.Time.

2. Rename Span to RawSpan.

Freeing up the name for the full tracing implementation where Span is an
abstraction instead of the wire format.

While this is theoretically a breaking change, it is one that should not
affect users, as those names were only used internally by the Sentry
OpenTelemetry Exporter (we can update it) and are marked as experimental
APIs.