doc: Remove deprecation of sentryhttp.HandleFunc
Reverts the deprecation note added in
36c74322d86a295bf2cde8854576e42a77526bb3.

While sketching out tracing support for the sentryhttp package, I
realized that the deprecation was not necessary, since
sentryhttp.HandleFunc is particularly useful in one situation, and now
we have a more explicit example of that usage.

Typically, user code has either an http.Handler (http.HandlerFunc
implements http.Handler) or a function literal of type
func(http.ResponseWriter, *http.Request) to pass as argument.

Because sentryhttp.Handle already takes an http.Handler parameter, what
is left for sentryhttp.HandleFunc to be useful at is taking in a handler
function literal, implicitly converting it to an http.HandlerFunc so
that users don't have to write the explicit conversion, saving a bit of
noise.