build: Bump required version of Iris
This mitigates license concerns voiced by Sentry customers, while
promoting the use of a more up-to-date version of the web framework.

Iris v12.1.8 is the latest release, released on Feb 16, 2020.

The version bump eliminates LGPL-licensed github.com/juju/... packages
from the module dependency chain. Note that sentry-go never depended on
those packages, and they are only ever present in a built binary if
using sentry-go/iris together with an old version of Iris, brought in as
a web framework dependency.