ref: Report Breadcrumb.Timestamp in local time
The serialization format, RFC3339, includes time zone information, so
the change has no effect for machines. When debugging, local time can be
easier to parse by humans.

Note that prior to this change the timestamp was set by the SDK in UTC
but not enforced in all code paths. If users set the timestamp
themselves, the original location was preserved.

Update Breadcrumb.MarshalJSON to expand commentary on why and how we
customize the serialization of breadcrumbs.

Add more serialization tests.