fix: Transaction envelope must include event_id
Without the event ID in the envelope header, the server generates a new
one, such that the event ID generated by the SDK was being ignored.

While fixing that, some more changes:

- Include the optional 'length' in the envelope item header (as
preferred by the spec)

- Rename getEnvelopeFromBody -> transactionEnvelopeFromBody, because the
helper function is limited to transaction envelopes and not the general
case.

- Use a JSON encoder to encode the parts of the envelope. This is the
safer and more obvious implementation, though we might move back to
hand-writing the JSON output if performance suffers.

- Do not ignore errors from http.NewRequest.

- Use a bytes.Reader instead of bytes.Buffer where we only need reading
off a []byte.