doc: Point users to preferable or more commonly used APIs (#302)
While not extensive, the clarifications are meant to help people with
inline hints when they use the SDK in their favorite editor.

Changing a client's transport in unsupported because setting
Client.Transport does not configure the transport. While users could
theoretically call Transport.Configure themselves, the next problem is
that typically a single client is used by multiple hubs on concurrent
goroutines, and directly changing the transport without proper
syncronization is a data race.