Allow custom error types to emit more specific Stacktraces
This change creates a new interface, `StacktraceGetter`, which allows
custom error types to embed a more specific Stacktrace within
themselves to be caught by more generic error handlers, and updates
`ExtraceStacktrace` to rely on this if implemented.

In addition to more specific Stacktraces, this partially de-couples
embedded Stacktraces from their implementation and creates an opt-in
integration of third-party error packages, rather than expanding the
scope of reflection in `ExtraceStacktrace`. The `GetSentryStracktrace`
in `StacktraceGetter` explicitly references Sentry to ensure this
interface is opt-in and unlikely to conflict with existing Stacktrace
implementations.

This should also work with `Recover*` handlers without modification, in
the case where custom error types are being thrown.