fix: Send root span data as Event.Extra (#329)
* fix: Send root span data as Event.Extra

Pending spec clarification since implementation in Relay and different
SDKs diverged.

For now sending data in Event.Extra gets the information into the
Transaction view in Sentry.

* docs: move comment to test

The test explains the intent and desired behavior. If the desired
behavior changes, we need to update the test and are thus more likely to
update/remove the comment.