ci: Use race detector only once for PRs
We run tests on well-supported platforms, such that there is little
chance the race detector will detect a race in one operating system but
not in the others.

We can save resources, human and machine, by not using the race detector
too many times in every PR.

To be on the safe side, leave it on for pushes to master and release
branches.

Remove it from GOPATH Mode completely since it is redundant.