build: Bump github.com/labstack/echo/v4 to v4.5.0
Removes github.com/dgrijalva/jwt-go@v3.2.0 from go.sum.

The referenced module is affected by CVE-2020-26160, and some tools end
up reporting on it even when it is not part of the actual build list.

Bumping echo to v4.5.0 and not the current latest tagged version
(v4.6.1) makes tools happy while keeping support for a broader range of
recent releases.