docs: Remove migration guide
Keeping a link to the migration guide at the docs site.

This is in order to keep a single source of truth and simplify
maintenance.

The exact same migration guide was repeated in two different places, we
choose to keep the one in the docs site because it provides an unified
experience to users (compared to linking from the Sentry docs to a page
on GitHub or pkg.go.dev).