feat: Allow some non-Golang interfaces to be passed through in Events.
Some users may wish to use sentry-go as a mechanism to proxy non-Golang
events -- for instance, if reporting errors from a mobile or web
application.  We add Sentry event payload interfaces that make it easier to
support other platforms, including stack frame interface members for
"C-like" applications, and including the Debug Meta interface that allows
the Sentry backend to symbolicate application backtraces.

This change was authored under contract for FullStory.